fix: resolve all clippy lint warnings
- error.rs: restructure test_result_ok/err to avoid unnecessary_literal_unwrap
- llm.rs: convert loop+break to while let, collapse nested ifs
- bridge.rs: remove needless borrow on msg.chat()
- config_tests.rs: assert_eq!(x, true) → assert!(x)
- health.rs: move start_health_server before test module (items_after_test_module)
- outbound.rs: remove redundant closures in map/filter_map calls
- main.rs: remove unnecessary u8→u8 cast for max_connections

+++ rsworkspace/crates/telegram-bot/src/main.rs
@@ -341,7 +341,7 @@ async fn main() -> Result<()> {
             }
 
             if *max_connections > 0 {
-                options = options.max_connections(*max_connections as u8);
+                options = options.max_connections(*max_connections);
             }
 
             let listener = match teloxide::update_listeners::webhooks::axum(bot, options).await {
